Update README.md
Resume Analyzer is a web-based application designed to automate the process of screening and evaluating resumes. It uses Natural Language Processing (NLP) and machine learning to extract important details such as skills, education, and experience from uploaded resumes, and then scores them based on relevance to a given job description. The frontend is developed using Next.js with TypeScript for a smooth and responsive user experience, while the backend is built in Python to handle data processing and model predictions. This tool helps streamline recruitment by reducing manual effort and improving the accuracy of candidate shortlisting.